+++ rpxy-lib/src/hyper_ext/mod.rs
@@ -12,5 +12,7 @@ pub(crate) mod rt {
 #[allow(unused)]
 pub(crate) mod body {
   pub(crate) use super::body_incoming_like::IncomingLike;
-  pub(crate) use super::body_type::{BoundedStreamBody, BoxBody, RequestBody, ResponseBody, empty, full};
+  pub(crate) use super::body_type::{
+    BoundedStreamBody, BoxBody, LimitedBody, LimitedIncoming, RequestBody, ResponseBody, empty, full,
+  };
 }